requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pandas==2.3.1
openpyxl==3.1.4
xlrd==2.0.1
python-dotenv==1.0.1
google-generativeai==0.7.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-json-logger==2.0.7
pytest==8.2.2
httpx==0.27.0
respx==0.21.1